Auth0 client secret configurable in Container
The Auth0 client secret is now hardcoded in dev env
only. In a container environment Caddy will template
the env variable AUTH0_CLIENT_SECRET into the js file
when serving it.

A future enhancement would be to make the dev secret
configurable via a .env in the root of this project
so that it's possible for other orgs to run this
against their own auth0 account.

Signed-off-by: Elliot Murphy <elliot@elliotmurphy.com>