Refactor worker build to simplify go imports
Previously, the docker build context for worker only had
the github.com/kindlyops/mappamundi/worker source code
visible inside the build container. The keycloak package was
brought in via the vendor tree. This made for an awkard
two-step dependency handling process, where the version of
github.com/kindlyops/mappamundi/havenapi/keycloak package
that worker was compiling with was the version referenced
in Gopkg.lock rather than the version that existed in the
revision of the monorepo that the worker binary was building
from.

This adjusts the worker build process so that the whole
monorepo is visible in the worker build so that package
import of havenapi/keycloak don't need to be vendored.

Simpler updates, and simpler to reason about the version
of the keycloak package that worker built against.

Signed-off-by: Elliot Murphy <statik@users.noreply.github.com>