First working chunks of keycloak-gatekeeper.
When signed out, /unleash/ prompts for authentication.
When signed in with appropriate roles, /unleash/ does SSO
When a gatekeeper session exists (after visiting /unleash/),
signing out of the Elm WebUI does not do SSO logout, so the
front page is signed out but /unleash/ still has a session.

This could cause problems from an admin signing out and not
realizing there is still an active admin session.

A new Keycloak client has been added, named gatekeeper, for
use by the gatekeeper proxy. The existing havendev client
that is used by keycloak.js in the SPA is a public client,
and gatekeeper requires a confidential client.

Ultimately it would be nice to get rid of keycloak.js and
use gatekeeper for all our authentication. Figuring out if
the Elm WebUI is able to key off of the gatekeeper cookies
to determine username and session state is TBD.

Signed-off-by: Elliot Murphy <statik@users.noreply.github.com>