Rework processing of LogOut message.
Still a bit funky to deal with logout, but this is the
cleanest way I could come up with.

We also need a background timer to warn the user their session is
going to be logged out automatically, and need to update all
the places that might get 401 responses from gatekeeper to
generate the logoutCmds as well.

Signed-off-by: Elliot Murphy <statik@users.noreply.github.com>